src/notifications/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
python-dotenv
aio-pika
asyncio